organises project for submission and adds small readme
