+++ lib/frame.js
