+++ lib/scorecard2.js
@@ -1,6 +1,7 @@
-const TenthFrameSpecial = require('./tenthframe');
-const Frame = require('./frame');
-
+// const TenthFrameSpecial = require('./tenthframe');
+// const Frame = require('./frame');
+const TenthFrameSpecial = require('../lib/tenthframe');
+const Frame = require('../lib/frame');
 
 /*
 10 Pin Scoring Rules:
